Rename is-function to isFunction.
