+++ lib/ModelRoot.js
@@ -1,4 +1,4 @@
-var isFunction = require("./support/is-function");
+var isFunction = require("./support/isFunction");
 var hasOwn = require("./support/hasOwn");
 var ImmediateScheduler = require("./schedulers/ImmediateScheduler");
