+++ lib/response/IdempotentResponse.js
@@ -14,6 +14,7 @@ var isArray = Array.isArray;
 var isPathValue = require("./../support/is-path-value");
 var isJsonEnvelope = require("./../support/is-json-envelope");
 var isJsonGraphEnvelope = require("./../support/is-json-graph-envelope");
+var isFunction = require("./../support/isFunction");
 
 function IdempotentResponse(subscribe) {
     Observable.call(this, subscribe);
