+++ lib/response/SetResponse.js
@@ -8,6 +8,7 @@ var InvalidSourceError = require("./../errors/InvalidSourceError.js");
 var arrayMap = require("./../support/array-map");
 var spreadJSON = require("./../get/util/spreadJSON");
 var arrayFlatMap = require("./../support/array-flat-map");
+var isFunction = require("./../support/isFunction");
 var emptyArray = new Array(0);
 
 function SetResponse(subscribe) {
