+++ lib/response/updateProgressively.js
@@ -1,4 +1,4 @@
-var isFunction = require("falcor/support/is-function");
+var isFunction = require("falcor/support/isFunction");
 
 module.exports = function updateProgressively() {
     var source = this;
